move accessories to dedicated folder
