+++ src/assets/icons/static/accessories/Hat.tsx
