+++ src/assets/icons/static/accessories/Mittens.tsx
