+++ src/assets/icons/static/accessories/RainBoots.tsx
