+++ src/assets/icons/static/accessories/Scarf.tsx
