+++ src/assets/icons/static/accessories/Sunglasses.tsx
