+++ src/assets/icons/static/accessories/Sunscreen.tsx
