+++ src/assets/icons/static/accessories/Toque.tsx
