+++ src/assets/icons/static/accessories/Umbrella.tsx
